--- js/components/userFeedback/redux/constatnts.js
@@ -1,12 +1,0 @@
-const prefix = 'ISSUE_';
-
-export const constants = {
-    RESET_FORM: prefix + 'RESET_FORM',
-
-    SET_NAME: prefix + 'SET_NAME',
-    SET_EMAIL: prefix + 'SET_EMAIL',
-    SET_TITLE: prefix + 'SET_TITLE',
-    SET_DESCRIPTION: prefix + 'SET_DESCRIPTION',
-    SET_RESPONSE: prefix + 'SET_RESPONSE',
-    SET_IMAGE_SOURCE: prefix + 'SET_IMAGE_SOURCE'
-};
